+++ sdk/examples/console/util.go
@@ -1,0 +1,53 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/stellar/go/clients/horizonclient"
+	"github.com/stellar/go/keypair"
+	"github.com/stellar/go/txnbuild"
+)
+
+func getSeqNum(client horizonclient.ClientInterface, accountID string) (int64, error) {
+	account, err := client.AccountDetail(horizonclient.AccountRequest{AccountID: accountID})
+	if err != nil {
+		return 0, fmt.Errorf("getting account %s: %w", accountID, err)
+	}
+	seqNum, err := account.GetSequenceNumber()
+	if err != nil {
+		return 0, fmt.Errorf("getting sequence number of account %s: %w", accountID, err)
+	}
+	return seqNum, nil
+}
+
+func fund(client horizonclient.ClientInterface, networkPassphrase string, accountKey *keypair.FromAddress) error {
+	rootKey := keypair.Root(networkPassphrase)
+	root, err := client.AccountDetail(horizonclient.AccountRequest{AccountID: rootKey.Address()})
+	if err != nil {
+		return err
+	}
+	tx, err := txnbuild.NewTransaction(txnbuild.TransactionParams{
+		SourceAccount:        &root,
+		IncrementSequenceNum: true,
+		BaseFee:              txnbuild.MinBaseFee,
+		Timebounds:           txnbuild.NewTimeout(300),
+		Operations: []txnbuild.Operation{
+			&txnbuild.CreateAccount{
+				Destination: accountKey.Address(),
+				Amount:      "10000",
+			},
+		},
+	})
+	if err != nil {
+		return err
+	}
+	tx, err = tx.Sign(networkPassphrase, rootKey)
+	if err != nil {
+		return err
+	}
+	_, err = client.SubmitTransaction(tx)
+	if err != nil {
+		return err
+	}
+	return nil
+}
